reworked as per suggestions
